requirements.txt
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
google-cloud-bigquery>=3.0.0
pyyaml>=5.0.0
jupyter>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

